requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
colorgram.py==1.2.0
Pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.8.1.78
torch==2.3.1
torchvision==0.18.1
timm==1.0.3
gunicorn==22.0.0
transformers==4.41.2
accelerate==0.32.1
sentencepiece==0.2.0
huggingface-hub==0.23.1